pyproject.toml
[tool.poetry]
name = "wikicite"
version = "0.1.2"
description = "Create ready-to-use Wikipedia citations from a series of inputs"
authors = ["peaky76 <robertjamespeacock@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
click = "^8.1.3"
pyyaml = "^6.0"

[tool.poetry.scripts]
wikicite = "wikicite.wikicite:cite"
test_wikicite = "tests.test_wikicite:runner"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
